added pyright and fixed a few bits
one day this will pass on strict...
