+++ apollo.py
@@ -110,7 +110,8 @@ async def sync(ctx: Context) -> None:
 
 
 @bot.event
-async def on_command_error(ctx: Context, error: Exception):
+async def on_command_error(ctx: Context, error: commands.CommandError):
+    assert ctx.command
     # await ctx.message.add_reaction("🚫")
     message = ""
     reraise = None
@@ -127,7 +128,7 @@ async def on_command_error(ctx: Context, error: Exception):
         message = f"Bot does not have permissions to do this. {str(error.text)}"
         reraise = error
     elif hasattr(error, "original"):
-        await on_command_error(ctx, error.original)
+        await on_command_error(ctx, error.original)  #  type: ignore
         return
     elif isinstance(error, errors.CommandError):
         message = str(error)
